rename JsonWriter as ArchesFileWriter and move from archesjson.py into archesfile.py, re #3313
